requirements.txt
asgiref==3.2.10
Django==3.1.13
django-crispy-forms==1.9.2
pytz==2020.1
sqlparse==0.4.2
djangorestframework
gunicorn
django-heroku
Pillow==9.5.0
pytesseract==0.3.10
opencv-python==4.8.0.76
numpy==1.24.3
scikit-learn==1.3.0